requirements.txt
python-telegram-bot==20.5
wikipedia-api==0.5.8
googletrans==4.0.0rc1
requests==2.31.0
